+++ OpenAI.SDK/ObjectModels/SharedModels/HyperParametersResponse.cs
@@ -7,11 +7,11 @@ namespace OpenAI.GPT3.ObjectModels.SharedModels
         [JsonPropertyName("batch_size")] public int? BatchSize { get; set; }
 
         [JsonPropertyName("learning_rate_multiplier")]
-        public float? LearningRateMultiplier { get; set; }
+        public double? LearningRateMultiplier { get; set; }
 
         [JsonPropertyName("n_epochs")] public int? NEpochs { get; set; }
 
         [JsonPropertyName("prompt_loss_weight")]
-        public float? PromptLossWeight { get; set; }
+        public double? PromptLossWeight { get; set; }
     }
 }
