+++ OpenAI.SDK/ObjectModels/SharedModels/IOpenAiModels.cs
@@ -19,7 +19,7 @@
 
         public interface ITemperature
         {
-            float? Temperature { get; set; }
+            double? Temperature { get; set; }
         }
 
         public interface ICreatedAt
